+++ misc/metrics/CHANGELOG.md
@@ -1,3 +1,11 @@
-## Version 0.1.0 [2021-11-01]
+# 0.2.0 [2021-11-16]
+
+- Include gossipsub metrics (see [PR 2316]).
+
+- Update dependencies.
+
+[PR 2316]: https://github.com/libp2p/rust-libp2p/pull/2316
+
+# 0.1.0 [2021-11-01]
 
 - Add initial version.