+++ protocols/gossipsub/CHANGELOG.md
@@ -1,11 +1,16 @@
-# 0.33.1 [unreleased]
+# 0.34.0 [2021-11-16]
+
+- Add topic and mesh metrics (see [PR 2316]).
 
 - Fix bug in internal peer's topics tracking (see [PR 2325]).
 
 - Use `instant` and `futures-timer` instead of `wasm-timer` (see [PR 2245]).
 
+- Update dependencies.
+
 [PR 2245]: https://github.com/libp2p/rust-libp2p/pull/2245
 [PR 2325]: https://github.com/libp2p/rust-libp2p/pull/2325
+[PR 2316]: https://github.com/libp2p/rust-libp2p/pull/2316
 
 # 0.33.0 [2021-11-01]